Cleaned up Home.tsx
