+++ src/apis/UTXOManager/UTXOManager.ts
@@ -4,6 +4,7 @@ import DatabaseService from '../DatabaseManager/DatabaseService';
 import ElectrumService from '../ElectrumServer/ElectrumServer';
 import { Network } from '../../redux/networkSlice';
 import { store } from '../../redux/store';
+import { setUTXOs } from '../../redux/utxoSlice';
 
 export default async function UTXOManager() {
   const dbService = DatabaseService();
@@ -20,6 +21,7 @@ export default async function UTXOManager() {
     fetchUTXOsByAddress,
     checkNewUTXOs,
     deleteUTXOs,
+    fetchUTXOs,
   };
 
   async function storeUTXOs(utxos) {
@@ -211,5 +213,77 @@ export default async function UTXOManager() {
         );
       }
     }
+  }
+  async function fetchUTXOs(walletKeys) {
+    // setFetchingUTXOs(true);
+    const utxosMap = {};
+    const cashTokenUtxosMap = {};
+    const uniqueUTXOs = new Set();
+    const loadingState = {};
+
+    const electrumService = ElectrumService();
+    await electrumService.electrumConnect();
+
+    for (const key of walletKeys) {
+      loadingState[key.address] = true;
+    }
+    // setLoading(loadingState);
+
+    for (let i = 0; i < walletKeys.length; i++) {
+      const key = walletKeys[i];
+      console.log(`Key ${i}:`, key.address);
+      const addressUTXOs = await fetchUTXOsByAddress(
+        state.wallet_id.currentWalletId,
+        key.address
+      );
+      console.log(`Fetched UTXOs for address ${key.address}:`, addressUTXOs);
+
+      utxosMap[key.address] = addressUTXOs.filter((utxo) => {
+        const utxoKey = `${utxo.tx_hash}-${utxo.tx_pos}-${utxo.address}`;
+        if (uniqueUTXOs.has(utxoKey)) {
+          return false;
+        }
+        if (utxo.address === key.address && !utxo.token_data) {
+          uniqueUTXOs.add(utxoKey);
+          total += utxo.amount;
+          return true;
+        }
+        return false;
+      });
+
+      cashTokenUtxosMap[key.address] = addressUTXOs.filter((utxo) => {
+        const utxoKey = `${utxo.tx_hash}-${utxo.tx_pos}-${utxo.address}`;
+        if (uniqueUTXOs.has(utxoKey)) {
+          return false;
+        }
+        if (utxo.address === key.address && utxo.token_data) {
+          uniqueUTXOs.add(utxoKey);
+          total += utxo.amount;
+          return true;
+        }
+        return false;
+      });
+
+      loadingState[key.address] = false;
+      // setLoading({ ...loadingState });
+      // setUtxoProgress(((i + 1) / walletKeys.length) * 100);
+    }
+
+    await electrumService.electrumDisconnect();
+
+    // Update Redux with the new UTXOs
+    const newUTXOs = walletKeys.reduce((acc, key) => {
+      acc[key.address] = [
+        ...utxosMap[key.address],
+        ...cashTokenUtxosMap[key.address],
+      ];
+      return acc;
+    }, {});
+    store.dispatch(setUTXOs({ newUTXOs }));
+
+    return { utxosMap, cashTokenUtxosMap };
+    // setUtxos(utxosMap);
+    // setCashTokenUtxos(cashTokenUtxosMap);
+    // setFetchingUTXOs(false);
   }
 }
